Add example and documentation for BSPM evaluation
