+++ docs/source/auto_docs/mach_eval.analyzers.electromagnetic.rst
