+++ docs/source/auto_docs/mach_eval.rst
