+++ docs/source/auto_docs/mach_opt.rst
